+++ Lib/fontbakery/profiles/os2.py
@@ -167,17 +167,21 @@ def com_google_fonts_check_xavgcharwidth(ttFont):
     if current_value == expected_value or difference == 1:
         yield PASS, "OS/2 xAvgCharWidth value is correct."
     elif difference < ACCEPTABLE_ERROR:
-        yield INFO, (f"OS/2 xAvgCharWidth is {current_value} but it should be"
-                     f" {expected_value} which corresponds to {calculation_rule}."
-                     f" These are similar values, which"
-                     f" may be a symptom of the slightly different"
-                     f" calculation of the xAvgCharWidth value in"
-                     f" font editors. There's further discussion on"
-                     f" this at https://github.com/googlefonts/fontbakery"
-                     f"/issues/1622")
+        yield INFO, \
+              Message("xAvgCharWidth-close",
+                      f"OS/2 xAvgCharWidth is {current_value} but it should be"
+                      f" {expected_value} which corresponds to {calculation_rule}."
+                      f" These are similar values, which"
+                      f" may be a symptom of the slightly different"
+                      f" calculation of the xAvgCharWidth value in"
+                      f" font editors. There's further discussion on"
+                      f" this at https://github.com/googlefonts/fontbakery"
+                      f"/issues/1622")
     else:
-        yield WARN, (f"OS/2 xAvgCharWidth is {current_value} but it should be"
-                     f" {expected_value} which corresponds to {calculation_rule}.")
+        yield WARN, \
+              Message("xAvgCharWidth-wrong",
+                      f"OS/2 xAvgCharWidth is {current_value} but it should be"
+                      f" {expected_value} which corresponds to {calculation_rule}.")
 
 
 @check(
@@ -208,14 +212,18 @@ def com_adobe_fonts_check_fsselection_matches_macstyle(ttFont):
     os2_bold = (ttFont['OS/2'].fsSelection & FsSelection.BOLD) != 0
     if head_bold != os2_bold:
         failed = True
-        yield FAIL, ("The OS/2.fsSelection and head.macStyle "
-                     "bold settings do not match.") # FIXME: Needs a keyword
+        yield FAIL, \
+              Message("fsselection-macstyle-bold",
+                      "The OS/2.fsSelection and head.macStyle " \
+                      "bold settings do not match.")
     head_italic = (ttFont['head'].macStyle & MacStyle.ITALIC) != 0
     os2_italic = (ttFont['OS/2'].fsSelection & FsSelection.ITALIC) != 0
     if head_italic != os2_italic:
         failed = True
-        yield FAIL, ("The OS/2.fsSelection and head.macStyle "
-                     "italic settings do not match.") # FIXME: Needs a keyword
+        yield FAIL, \
+              Message("fsselection-macstyle-italic",
+                      "The OS/2.fsSelection and head.macStyle " \
+                      "italic settings do not match.")
     if not failed:
         yield PASS, ("The OS/2.fsSelection and head.macStyle "
                      "bold and italic settings match.")
@@ -260,10 +268,12 @@ def com_adobe_fonts_check_family_bold_italic_unique_for_nameid1(RIBBI_ttFonts):
     for (family_name, bold_italic), count in counter.items():
         if count > 1:
             failed = True
-            yield FAIL, (f"Family '{family_name}' has {count} fonts"
-                         f" (should be no more than 1) with the"
-                         f" same OS/2.fsSelection bold & italic settings:"
-                         f" {bold_italic}")
+            yield FAIL, \
+                  Message("unique-fsselection",
+                          f"Family '{family_name}' has {count} fonts"
+                          f" (should be no more than 1) with the"
+                          f" same OS/2.fsSelection bold & italic settings:"
+                          f" {bold_italic}")
     if not failed:
         yield PASS, ("The OS/2.fsSelection bold & italic settings were unique "
                      "within each compatible family group.")
@@ -294,7 +304,9 @@ def com_google_fonts_check_code_pages(ttFont):
        not hasattr(ttFont['OS/2'], "ulCodePageRange2") or \
        (ttFont['OS/2'].ulCodePageRange1 == 0 and \
         ttFont['OS/2'].ulCodePageRange2 == 0):
-        yield FAIL, ("No code pages defined in the OS/2 table"
-                     " ulCodePageRange1 and CodePageRange2 fields.")
+        yield FAIL, \
+              Message("no-code-pages",
+                      "No code pages defined in the OS/2 table"
+                      " ulCodePageRange1 and CodePageRange2 fields.")
     else:
         yield PASS, "At least one code page is defined."